requirements.txt
clickhouse-driver>=0.2.0
kafka-python>=2.0.2
pandas>=1.3.0
mysql-connector-python>=8.0.0
pymysql>=1.0.2
pyspark>=3.3.0
